requirements.txt
google-cloud-datastore
cloud-files
codecov
